backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
supabase==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
pydantic-settings==2.1.0

# Mobile optimizations
redis==5.0.1
slowapi==0.1.9
python-jose[cryptography]==3.3.0

# Performance monitoring
prometheus-client==0.19.0 